backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
playwright==1.42.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2
pydantic==2.6.3 